Add launch support with Mixins
This requires both the `org.ow2.asm:asm-util:6.2` and `org.spongepowered:mixin:0.8-SNAPSHOT` libraries to be present on the game's classpath during startup. As
of today, this is not possible in vanilla Forge without modifying the game's
version manifest to include these two libraries.

As such, this will only work for people really willing to tinker with the game.